chat_app/requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
mysql-connector-python==9.2.0
openai==1.76.2
Requests==2.32.3
